+++ mov/src/org/mov/quote/MetaStock2QuoteFilter.java
@@ -93,13 +93,13 @@ public class MetaStock2QuoteFilter implements EODQuoteFilter {
                 }
 
                 try {
-                    float day_open = Float.parseFloat(quoteParts[i++]);
-                    float day_high = Float.parseFloat(quoteParts[i++]);
-                    float day_low = Float.parseFloat(quoteParts[i++]);
-                    float day_close = Float.parseFloat(quoteParts[i++]);
+                    double day_open = Double.parseDouble(quoteParts[i++]);
+                    double day_high = Double.parseDouble(quoteParts[i++]);
+                    double day_low = Double.parseDouble(quoteParts[i++]);
+                    double day_close = Double.parseDouble(quoteParts[i++]);
                     
                     // Convert volume from 1/100th volume to real volume
-                    int day_volume = Integer.parseInt(quoteParts[i++]) * 100;
+                    long day_volume = Long.parseLong(quoteParts[i++]) * 100;
                     quote = new EODQuote(symbol, date, day_volume, day_low, day_high,
                                          day_open, day_close);
                 }
